requirements.txt
aniso8601==9.0.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.2
daphne==4.1.0
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.57.2
django-cors-headers==4.3.1
django-filter==23.5
django-ipware==6.0.4
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
Faker==23.2.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Markdown==3.5.2
msgpack==1.0.7
mysqlclient==2.2.4
oauthlib==3.2.2
phonenumbers==8.13.30
pillow==10.2.0
promise==2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-ipware==2.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
text-unidecode==1.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.0
watchgod==0.8.2
whitenoise==6.6.0
zope.interface==6.2
